requirements.txt
mysql-connector-python
fpdf
num2words